refactor #97: 쿠키 설정 Getter와 응답 코드 수정
