+++ user-service/src/main/java/com/musinsam/userservice/app/domain/auth/vo/AuthErrorCode.java
@@ -1,4 +1,4 @@
-package com.musinsam.userservice.app.global.response;
+package com.musinsam.userservice.app.domain.auth.vo;
 
 import com.musinsam.common.exception.ErrorCode;
 import lombok.Getter;
